refactor: move insee api connector to a separate file
